// extensions/mite/src/model/Project.tsx
export interface Project {
  budget: 0;
  budget_type: string;
  created_at: string;
  customer_id: number;
  hourly_rate: number;
  id: number;
  name: string;
  note: string;
  updated_at: string;
  archived: boolean;
  active_hourly_rate: null;
  hourly_rates_per_service: [];
  customer_name: string;
}